+++ Bots/marks_coding_corner/ConsPrinter.py
@@ -3,14 +3,14 @@ import os
 from Py4GWCoreLib import GLOBAL_CACHE
 from Py4GWCoreLib import Bags
 from Py4GWCoreLib import Botting
+from Py4GWCoreLib import ConsoleLog
 from Py4GWCoreLib import Item
 from Py4GWCoreLib import ItemArray
 from Py4GWCoreLib import ModelID
 from Py4GWCoreLib import Py4GW
+from Py4GWCoreLib import PyImGui
 from Py4GWCoreLib import Routines
 from Py4GWCoreLib import Trading
-from Py4GWCoreLib import ConsoleLog
-from Py4GWCoreLib import PyImGui
 
 try:
     script_directory = os.path.dirname(os.path.abspath(__file__))
