+++ Bots/marks_coding_corner/DervFeatherFarm.py
@@ -2,11 +2,15 @@ import random
 
 from Bots.marks_coding_corner.utils.loot_utils import VIABLE_LOOT
 from Bots.marks_coding_corner.utils.loot_utils import get_valid_loot_array
-from Bots.marks_coding_corner.utils.loot_utils import sell_non_essential_mats
+from Bots.marks_coding_corner.utils.loot_utils import identify_and_salvage_items
+from Bots.marks_coding_corner.utils.merch_utils import buy_id_kits
+from Bots.marks_coding_corner.utils.merch_utils import buy_salvage_kits
+from Bots.marks_coding_corner.utils.merch_utils import sell_non_essential_mats
+from Bots.marks_coding_corner.utils.town_utils import return_to_outpost
+from Py4GWCoreLib import *
 from Py4GWCoreLib.Builds.DervFeatherFarmer import SENSALI_MODEL_IDS
 from Py4GWCoreLib.Builds.DervFeatherFarmer import DervBuildFarmStatus
 from Py4GWCoreLib.Builds.DervFeatherFarmer import DervFeatherFarmer
-from Py4GWCoreLib import *
 
 try:
     script_directory = os.path.dirname(os.path.abspath(__file__))
@@ -65,24 +69,6 @@ is_looting = False
 
 
 # region Direct Bot Actions
-def return_to_outpost():
-    if GLOBAL_CACHE.Map.IsOutpost():
-        return
-
-    is_explorable = GLOBAL_CACHE.Map.IsExplorable()
-    while is_explorable:
-        is_map_ready = GLOBAL_CACHE.Map.IsMapReady()
-        is_party_loaded = GLOBAL_CACHE.Party.IsPartyLoaded()
-        is_party_defeated = GLOBAL_CACHE.Party.IsPartyDefeated() or GLOBAL_CACHE.Player.GetMorale() < 100
-
-        if is_map_ready and is_party_loaded and is_explorable and is_party_defeated:
-            yield from Routines.Yield.Player.Resign()
-            yield from Routines.Yield.wait(2000)
-            GLOBAL_CACHE.Party.ReturnToOutpost()
-        else:
-            is_explorable = GLOBAL_CACHE.Map.IsExplorable()
-            yield from Routines.Yield.wait(4000)
-
 
 def load_skill_bar(bot: Botting):
     yield from bot.config.build_handler.LoadSkillBar()
@@ -188,25 +174,6 @@ def loot_items():
         item_id_blacklist = item_id_blacklist + failed_items_id
     ConsoleLog(FEATHER_FARMER, 'Looting items finished')
     yield
-
-
-def identify_and_salvage_items():
-    yield from Routines.Yield.wait(1500)
-    yield from AutoInventoryHandler().IDAndSalvageItems()
-
-
-def buy_id_kits():
-    yield from Routines.Yield.wait(1500)
-    kits_in_inv = GLOBAL_CACHE.Inventory.GetModelCount(ModelID.Identification_Kit)
-    if kits_in_inv < 3:
-        yield from Routines.Yield.Merchant.BuyIDKits(3)
-
-
-def buy_salvage_kits():
-    yield from Routines.Yield.wait(1500)
-    kits_in_inv = GLOBAL_CACHE.Inventory.GetModelCount(ModelID.Salvage_Kit)
-    if kits_in_inv <= 3:
-        yield from Routines.Yield.Merchant.BuySalvageKits(10)
 
 
 # endregion
