+++ Bots/marks_coding_corner/VaettirMarksMods.py
@@ -1,17 +1,23 @@
+from typing import List
+from typing import Tuple
+
 import Py4GW
 from Bots.marks_coding_corner.utils.loot_utils import VIABLE_LOOT
-from Bots.marks_coding_corner.utils.loot_utils import sell_non_essential_mats
 from Bots.marks_coding_corner.utils.loot_utils import get_valid_salvagable_loot_array
-from Py4GWCoreLib import Routines, Botting, ActionQueueManager, ConsoleLog, GLOBAL_CACHE, Utils
+from Bots.marks_coding_corner.utils.merch_utils import sell_non_essential_mats
+from Py4GWCoreLib import GLOBAL_CACHE
+from Py4GWCoreLib import ActionQueueManager
+from Py4GWCoreLib import Botting
+from Py4GWCoreLib import ConsoleLog
+from Py4GWCoreLib import Routines
 from Py4GWCoreLib import ThrottledTimer
-from Py4GWCoreLib.enums import ModelID, Range, TitleID
-
+from Py4GWCoreLib import Utils
 from Py4GWCoreLib.BuildMgr import BuildMgr
-
-from Py4GWCoreLib.Builds import SF_Mes_vaettir
 from Py4GWCoreLib.Builds import SF_Ass_vaettir
-
-from typing import List, Tuple
+from Py4GWCoreLib.Builds import SF_Mes_vaettir
+from Py4GWCoreLib.enums import ModelID
+from Py4GWCoreLib.enums import Range
+from Py4GWCoreLib.enums import TitleID
 
 VAETTIR_FARM_BY_MARK = "Vaettir By Mark"
 HANDLE_STUCK_JAGA_MORAINE = "HandleStuckJagaMoraine"
