+++ Bots/marks_coding_corner/utils/merch_utils.py
@@ -1,0 +1,40 @@
+from Py4GWCoreLib import GLOBAL_CACHE
+from Py4GWCoreLib import Bags
+from Py4GWCoreLib import ItemArray
+from Py4GWCoreLib import ModelID
+from Py4GWCoreLib import Routines
+
+
+def sell_non_essential_mats():
+    MERCHABLE_CRAFTING_MATERIALS_MODEL_ID = [
+        ModelID.Wood_Plank,
+        ModelID.Scale,
+        ModelID.Tanned_Hide_Square,
+        ModelID.Bolt_Of_Cloth,
+        ModelID.Granite_Slab,
+        ModelID.Chitin_Fragment,
+    ]
+    bag_list = ItemArray.CreateBagList(Bags.Backpack, Bags.BeltPouch, Bags.Bag1, Bags.Bag2)
+    all_items = ItemArray.GetItemArray(bag_list)
+    item_ids_to_sell = []
+
+    for item_id in all_items:
+        if GLOBAL_CACHE.Item.GetModelID(item_id) in MERCHABLE_CRAFTING_MATERIALS_MODEL_ID:
+            item_ids_to_sell.append(item_id)
+
+    yield from Routines.Yield.Merchant.SellItems(item_ids_to_sell)
+
+
+def buy_id_kits():
+    yield from Routines.Yield.wait(1500)
+    kits_in_inv = GLOBAL_CACHE.Inventory.GetModelCount(ModelID.Identification_Kit)
+    sup_kits_in_inv = GLOBAL_CACHE.Inventory.GetModelCount(ModelID.Superior_Identification_Kit)
+    if (kits_in_inv + sup_kits_in_inv) < 1:
+        yield from Routines.Yield.Merchant.BuyIDKits(1)
+
+
+def buy_salvage_kits():
+    yield from Routines.Yield.wait(1500)
+    kits_in_inv = GLOBAL_CACHE.Inventory.GetModelCount(ModelID.Salvage_Kit)
+    if kits_in_inv < 2:
+        yield from Routines.Yield.Merchant.BuySalvageKits(3)
